tcp/udp: Change port type to uint16_t and timeout type to int64_t

+++ netlib/example/tcp_server_example.cpp
@@ -6,7 +6,7 @@
 
 using namespace net;
 
-static const int kPort = 8080;
+static const uint16_t kPort = 8080;
 
 static void handle(const std::shared_ptr<TCPSocket>& clientsock) {
     std::cout << "ThreadID = " << std::this_thread::get_id() << std::endl;
